fix : 폴더구조 확정된 내용으로 변경 및 초기 파일 셋팅
