+++ src/app/article/page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function ArticlePage() {
+  return (
+    <div>
+      <p>ArticlePage</p>
+    </div>
+  );
+}
