--- src/app/home/page.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-import { Button } from "@/components/ui/button";
-
-export default function Home() {
-  return (
-    <div>
-      <p>home</p>
-      <p>home2</p>
-      <Button>ㅋㅋ</Button>
-    </div>
-  );
-}
