+++ src/app/quiz/page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function QuizPage() {
+  return (
+    <div>
+      <p>QuizPage</p>
+    </div>
+  );
+}
