+++ src/app/workbook/page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function WorkBookPage() {
+  return (
+    <div>
+      <p>WorkBookPage</p>
+    </div>
+  );
+}
