+++ src/article/remotes/index.ts
@@ -1,0 +1,1 @@
+// api 통신관련
