+++ src/quiz/components/articlePopup/articlePopup.test.ts
@@ -1,0 +1,5 @@
+describe("FEW 단위 테스트", () => {
+  it("FEW 최고!", () => {
+    // ...
+  });
+});
