+++ src/quiz/types/index.ts
@@ -1,0 +1,1 @@
+//type
