+++ src/shared/constants/endpoint.ts
@@ -1,0 +1,1 @@
+export const LOGOUT_URL = "/api/v1/auth/logout";
