+++ src/workbook/components/articlePopup/index.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function articlePopup() {
+  return <div></div>;
+}
